pyproject.toml
[project]
name = "sec-watcher"
version = "0.1.0"
description = "Email notification system for SEC filings"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "schedule==1.2.2",
]
